Greatly reduced simulator file size through refactoring
