+++ Simulator/sim/__init__.py
@@ -1,2 +1,4 @@
 from .simulator_utility import simulator_utility
 from .simulator_core import simulator_core
+from .GUIManager import GUIManager
+from .simulation_draw import SimulationDraw
